feat: Add DownstreamApiBase class for handling downstream API calls
